Memoize problems to the database
Any problems that have status CLOSED are considered completed and are
saved to the database. In the future, they are pulled from there instead
of from the Dynatrace API.

If a problem in the problem feed has a root cause detected, problem details
are requested. This comes from the API or the database as described above