Expand problem summary for many problems
Extracted features are:
- Most heavily affected application
- A particular entity had many of the root causes
- A spike in problems during a time range

Feature extraction generally considers a feature interesting if it
deviates from the rest of the data by 2 standard deviations. In the
case of the time range, the time range asked for is split into 100
buckets of time and a single standard deviation is considered interesting
where 2 standard deviations is considered unusually strong.